run key-val from test. refactor Mock module to re-use more code.
