+++ Thundermint/Mock/KeyVal.hs
@@ -1,18 +1,43 @@
+{-# LANGUAGE DataKinds         #-}
+{-# LANGUAGE LambdaCase        #-}
 {-# LANGUAGE OverloadedStrings #-}
+{-# LANGUAGE RecordWildCards   #-}
+{-# LANGUAGE TupleSections     #-}
+
 -- |
 module Thundermint.Mock.KeyVal (
     genesisBlock
   , transitions
+  , executeSpec
   ) where
 
 import Control.Monad
-import Data.Map             (Map)
-import qualified Data.Map               as Map
+import Control.Monad.Catch
+import Control.Monad.IO.Class
+import Data.Int
+import Data.List
 
-import Thundermint.Consensus.Types
+import Data.Map        (Map)
+import Data.Maybe      (isJust)
+import System.FilePath ((</>))
+
+import qualified Data.Map as Map
+
+import Thundermint.Mock.Store  (newBlockStorage)
+import Thundermint.P2P.Network (createMockNode)
+import Thundermint.P2P.Network (newMockNet)
+
+import Thundermint.Blockchain.App
 import Thundermint.Blockchain.Interpretation
+import Thundermint.Blockchain.Types
+import Thundermint.Consensus.Types
+import Thundermint.Control
 import Thundermint.Crypto.Ed25519
-
+import Thundermint.Logger
+import Thundermint.Mock.KeyList
+import Thundermint.Mock.Types
+import Thundermint.P2P
+import Thundermint.Store
 ----------------------------------------------------------------
 --
 ----------------------------------------------------------------
@@ -45,3 +70,86 @@ transitions = BlockFold
     process (k,v) m
       | k `Map.member` m = Nothing
       | otherwise        = Just $ Map.insert k v m
+
+
+
+-------------------------------------------------------------------------------
+--
+-------------------------------------------------------------------------------
+
+interpretSpec
+  :: Maybe Int64                -- ^ Maximum height
+  -> FilePath
+  -> NetSpec                    --
+  -> IO [(BlockStorage 'RO IO Ed25519_SHA512 [(String,Int)], IO ())]
+interpretSpec maxH prefix NetSpec{..} = do
+  net <- newMockNet
+  forM (Map.toList netAddresses) $ \(addr, NodeSpec{..}) -> do
+    -- Prepare logging
+    let loggers = [ makeScribe s { scribe'path = fmap (prefix </>) (scribe'path s) }
+                  | s <- nspecLogFile
+                  ]
+    -- Create storage
+    storage  <- newBlockStorage prefix nspecDbName genesisBlock validatorSet
+    hChain   <- blockchainHeight storage
+    --
+    withLogEnv "TM" "DEV" loggers $ \logenv -> runLoggerT "general" logenv $ do
+      -- Blockchain state
+      bchState <- newBChState transitions
+                $ makeReadOnly (hoistBlockStorageRW liftIO storage)
+      _        <- stateAtH bchState (next hChain)
+      let appState = AppState
+            { appStorage        = hoistBlockStorageRW liftIO storage
+            --
+            , appValidationFun  = \h a -> do
+                st <- stateAtH bchState h
+                return $ isJust $ processBlock transitions h a st
+            --
+            , appBlockGenerator = \h -> case nspecByzantine of
+                Just "InvalidBlock" -> do
+                  return [("XXX", 123)]
+                _ -> do
+                  st <- stateAtH bchState h
+                  let Just k = find (`Map.notMember` st) ["K_" ++ show (n :: Int) | n <- [1 ..]]
+                  return [(k, addr)]
+            --
+            , appCommitCallback = \case
+                h | Just h' <- maxH
+                  , h > Height h'   -> throwM Abort
+                  | otherwise       -> return ()
+            , appValidator        = nspecPrivKey
+            , appNextValidatorSet = \_ _ -> return validatorSet
+            }
+      appCh <- newAppChans
+      return ( makeReadOnly storage
+             , runLoggerT "general" logenv $ runConcurrently
+                 [ setNamespace "net"
+                   $ startPeerDispatcher
+                       defCfg
+                       (createMockNode net "50000" addr)
+                       (addr,"50000")
+                       (map (,"50000") $ connections netAddresses addr)
+                       appCh
+                       (hoistBlockStorageRO liftIO $ makeReadOnly storage)
+                       nullMempool
+                 , setNamespace "consensus"
+                   $ runApplication defCfg appState appCh
+                 ]
+             )
+  where
+    netAddresses = Map.fromList $ [0::Int ..] `zip` netNodeList
+    connections  = case netTopology of
+      Ring    -> connectRing
+      All2All -> connectAll2All
+    validatorSet = makeValidatorSetFromPriv [ pk | Just pk <- nspecPrivKey <$> netNodeList ]
+
+
+executeSpec
+  :: Maybe Int64                -- ^ Maximum height
+  -> FilePath
+  -> NetSpec                    --
+  -> IO [BlockStorage 'RO IO Ed25519_SHA512 [(String,Int)]]
+executeSpec maxH prefix spec = do
+  actions <- interpretSpec maxH prefix spec
+  runConcurrently (snd <$> actions) `catch` (\Abort -> return ())
+  return $ fst <$> actions
