+++ spec/keyval-stm.json
@@ -2,12 +2,13 @@
   "netNodeList" :
   [
     { "nspecPrivKey"     : "2K7bFuJXxKf5LqogvVRQjms2W26ZrjpvUjo5LdvPFa5Y",
-      "nspecLogFile"     : 
+      "nspecLogFile"     :
       [{ "type"      : "ScribeJSON",
          "path"      : "logs/node-1",
          "severity"  : "Debug",
          "verbosity" : "V2"
-      }]
+       }],
+        "nspecWalletKeys"  : [0,0]
     },
     { "nspecPrivKey"     : "4NSWtMsEPgfTK25tCPWqNzVVze1dgMwcUFwS5WkSpjJL",
       "nspecByzantine"   : "InvalidBlock",
@@ -16,7 +17,8 @@
          "path"      : "logs/node-2",
          "severity"  : "Debug",
          "verbosity" : "V2"
-      }]
+       }],
+        "nspecWalletKeys"  : [0,0]
     },
     { "nspecPrivKey"     : "3Fj8bZjKc53F2a87sQaFkrDas2d9gjzK57FmQwnNnSHS",
       "nspecLogFile"     :
@@ -24,7 +26,8 @@
          "path"      : "logs/node-3",
          "severity"  : "Debug",
          "verbosity" : "V2"
-      }]
+       }],
+        "nspecWalletKeys"  : [0,0]
     },
     { "nspecPrivKey"     : "D2fpHM1JA8trshiUW8XPvspsapUvPqVzSofaK1MGRySd",
       "nspecLogFile"     :
@@ -32,8 +35,12 @@
          "path"      : "logs/node-4",
          "severity"  : "Debug",
          "verbosity" : "V2"
-      }]
+       }],
+        "nspecWalletKeys"  : [0,0]
     }
   ],
-  "netTopology"       : "All2All"
+  "netTopology"       : "All2All",
+  "netInitialDeposit" : 1000,
+  "netInitialKeys"    : 2000
+
 }
